Auth.js

+++ src/app/api/todos/route.ts
@@ -1,6 +1,7 @@
 import { NextResponse, NextRequest } from 'next/server'
 import prisma from '@/lib/prisma'
 import * as yup from 'yup';
+import { getUserServerSession } from '@/auth/actions/auth-actions';
 
 export async function GET(request: Request) {
     const { searchParams } = new URL(request.url);
@@ -34,9 +35,15 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
 
+    const user = await getUserServerSession();
+
+    if (!user) {
+        return NextResponse.json('No autorizado', { status: 401 });
+    }
+
     try {
         const { description, complete } = await postSchema.validate(await request.json());
-        const todo = await prisma.todo.create({ data: { description, complete } });
+        const todo = await prisma.todo.create({ data: { description, complete, userId: user.id } });
         return NextResponse.json(todo);
     } catch (error) {
         return NextResponse.json(error, { status: 400 });
@@ -46,8 +53,14 @@ export async function POST(request: Request) {
 
 export async function DELETE(request: Request) {
 
+    const user = await getUserServerSession();
+
+    if (!user) {
+        return NextResponse.json('No autorizado', { status: 401 });
+    }
+
     try {
-        await prisma.todo.deleteMany({ where: { complete: true }});
+        await prisma.todo.deleteMany({ where: { complete: true, userId: user.id }});
     
         return NextResponse.json({
             message: 'Todos eliminados correctamente'
